+++ webService/src/main/java/com/asklepios/hospitalreservation_asklepios/Controller/JoinController.java
@@ -1,27 +1,20 @@
 package com.asklepios.hospitalreservation_asklepios.Controller;
 
 import com.asklepios.hospitalreservation_asklepios.Service.IF_UserService;
-import com.asklepios.hospitalreservation_asklepios.Service.IM_UserService;
 import com.asklepios.hospitalreservation_asklepios.Util.Profile_ImageUtil;
 import com.asklepios.hospitalreservation_asklepios.VO.DoctorVO;
+import com.asklepios.hospitalreservation_asklepios.VO.MemberVO;
 import com.asklepios.hospitalreservation_asklepios.VO.UserVO;
-import jakarta.activation.DataSource;
-
-import jakarta.activation.FileDataSource;
-import org.apache.commons.io.FileUtils;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
-
 import org.springframework.ui.Model;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.File;
+import java.io.IOException;
 import java.util.UUID;
+
 
 @Controller
 public class JoinController {
@@ -78,4 +71,31 @@ public class JoinController {
         return "userJoin/successJoin";
     }
 
+    @GetMapping("/socialInfo")
+    public String socialInfo(Model model, HttpServletResponse response, HttpServletRequest request) throws IOException {
+        model.addAttribute("user", userService.findMember());
+        response.setContentType("text/html; charset=UTF-8");
+        response.setCharacterEncoding("UTF-8");
+        request.setCharacterEncoding("UTF-8");
+        response.getWriter().println("<script> alert('추가 정보 작성을 위해 해당 페이지로 이동합니다.');</script>");
+//        response.getWriter().close();
+        return "userJoin/insertSocialCommonInfo";
+    }
+    @PostMapping("/saveSocialInfo")
+    public String saveSocialInfo(@ModelAttribute("user") UserVO user
+            ,@RequestParam(value = "file", required = false) MultipartFile file ) throws Exception {
+        String newFileName = profileImageUtil.storeFile(file);
+        MemberVO member = userService.findMember();
+        user.setUser_authority("ROLE_scClient");
+        user.setUser_id(member.getUser_id());
+        user.setUser_password("Oauth2");
+        user.setUser_image(newFileName);
+        System.out.println(user.toString());
+        userService.modifySocialUserCommonInfo(user);
+        return "redirect:/home";
+    }
+    @GetMapping("/showAgreement")
+    public String showAgreement(){
+        return "userJoin/socialAgreement";
+    }
 }
