+++ webService/src/main/resources/templates/userJoin/insertSocialCommonInfo.html
@@ -254,7 +254,7 @@
                 </td>
                 <td>
                     <input type="text" id="user_info_agreement" name="user_info_agreement" style="text-align: center; font-weight: bold; color: red; width:49.5%" value="disagree" readonly >
-                    <input type="button" class="mx-auto bg-gray-50 border border-gray-300 text-gray-900 rounded-lg focus:ring-blue-500 focus:border-blue-500 mx-20 p-5 focus:ring-2 focus:outline-none" onclick="popUp()" value="약관보기" style="width: 49.5%">
+                    <input type="button" class="mx-auto bg-gray-50 border border-gray-300 text-gray-900 rounded-lg focus:ring-blue-500 focus:border-blue-500 mx-20 p-5 focus:ring-2 focus:outline-none" onclick="popUp()" value="약관동의" style="width: 49.5%">
                 </td>
             </tr>
             <tr>
